+++ app/templates/index.html
@@ -5,15 +5,15 @@
     <title>Document</title>
 </head>
 <body>
-    Hello <em>world</em>
+    Hello World! The index page was successfully loaded!
 
-    <form action="/test" method="POST" enctype="multipart/form-data">
+    <!-- <form action="/test" method="POST" enctype="multipart/form-data">
         <label for="name">Name:</label>
         <input type="text" name="name" id="name">
         <label for="pet-name">Pet:</label>
         <input type="text" name="pet-name" id="pet-name">
         <input type="file" name="pet-picture" id="pet-picture">
         <button type="submit">Submit</button>
-    </form>
+    </form> -->
 </body>
 </html>
