+++ run.py
@@ -1,0 +1,2 @@
+from app import app
+app.run()
